+++ lib/src/constants/regex_constants.dart
@@ -1,7 +1,11 @@
 class RegexConstans {
   static RegexConstans _instace;
   static RegexConstans get instance {
-    if (_instace == null) _instace = RegexConstans._init();
+    if (_instace != null) {
+      return _instace;
+    }
+
+    _instace = RegexConstans._init();
     return _instace;
   }