+++ lib/src/int_extension.dart
@@ -1,5 +1,5 @@
 import 'dart:math';
 
 extension IntagerExtension on int {
-  int get randomColorValue => Random().nextInt(17);
+  int get randomColorValue => Random().nextInt(this ?? 17);
 }