+++ test/extension/string_extension_test.dart
@@ -3,7 +3,7 @@ import 'package:kartal/kartal.dart';
 
 void main() {
   test('adds one to input values', () {
-    var fifteenLiras = "15";
+    var fifteenLiras = '15';
 
     expect(fifteenLiras.isNotNullOrNoEmpty, true);
   });